feat: separate connecting to database and start server
