+++ resources/cities/City.js
